* develop:
  Fix #700
  Need to review by Jason
  Fix #450
  Work in pending for #725
  Fix #727
  Fix #716
  Fix #721
  Fix #720
  Fix #722

Conflicts:
	src/app/controllers/auth.js

+++ src/app/services/auth.js
@@ -342,6 +342,7 @@ angular.module("proton.authentication", [
 
         // Removes all connection data
         logout: function() {
+            var accessToken = window.sessionStorage[CONSTANTS.OAUTH_KEY + ":AccessToken"];
             // Completely clear sessionstorage
             window.sessionStorage.clear();
 
@@ -351,9 +352,13 @@ angular.module("proton.authentication", [
             this.user = null;
 
             // HACKY ASS BUG
-            $http.delete(url.get() + "/auth").then( function() {
+            if(angular.isDefined(accessToken)) {
+                $http.delete(url.get() + "/auth").then( function() {
+                    location.reload();
+                });
+            } else {
                 location.reload();
-            });
+            }
 
             // THIS SHOULD BE RE-ENABLED WHEN WE FIX THE BUG
             // $rootScope.isLoggedIn = false;
